Added basic cli support
