+++ HCPinterface/cli/downloader.py
@@ -2,17 +2,25 @@
 
 # Downloads or deletes files from selected bucket on the HCP.
 
+import click
 import glob
 import argparse
 import os
 import json
 import sys
 import time
+
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from hcp.hcp import HCPManager
 from hci import hci
 
 ##############################################
+
+@click.group()
+def download():
+    """Downloads or deletes files from the selected bucket on the HCP"""
+    return
+
 def check(hcpm, args, pretty):
     if args.query:
         try:
@@ -26,8 +34,9 @@ def check(hcpm, args, pretty):
                 lst = string.replace('"','').replace("\\","").replace("[","").replace("]","").replace(";",",").split(",")
             print(f"Metadata file: {meta}")
             for i in lst:
-                if args.query in i:
+                if args.query in os.path.basename(i) or args.query in i:
                     print("check:",i)
+                    name = i.replace(".fastq.gz", ".fasterq").strip() # Replace suffix. 
 
         except:
             print(f"File(s) does not exists: {args.query}")
@@ -36,19 +45,26 @@ def check(hcpm, args, pretty):
         obj = hcpm.get_object(args.key) # Get object with key.
 
 # Download files using query, e.g. runid or sample name.
-def download(hcpm, args, pretty):
+def download_files(hcpm, args, pretty):
     if args.query:
         f = pretty
         results= f["results"]
         for item in results:
             itm = item["metadata"]
             samples = itm["samples_Fastq_paths"]
-        for i in samples:
-            obj = hcpm.get_object(i) # Get object with json.
-            if obj is not None:
-                hcpm.download_file(obj, f"{args.output}/"+os.path.basename(i)) # Downloads file.
-            else:
-                print(f"File: '{s}' does not exist in bucket '{args.bucket}' on the HCP")
+            string = "".join(samples).strip("[]").strip("{]}'")
+            lst = string.replace('"','').replace("\\","").replace("[","").replace("]","").replace(";",",").split(",")
+
+        for i in lst:
+            if args.query in os.path.basename(i) or args.query in i:
+                s = os.path.basename(i)
+                print("downloading:",s.replace(".fastq.gz", ".fasterq").strip())
+                name = s.replace(".fastq.gz", ".fasterq").strip() # Replace suffix. 
+                obj = hcpm.get_object(name) # Get object with json.
+                if obj is not None:
+                    hcpm.download_file(obj, f"{args.output}/"+os.path.basename(name)) # Downloads file.
+                else:
+                    print(f"File: '{name}' does not exist in bucket '{args.bucket}' on the HCP")
 
     else:
         obj = hcpm.get_object(args.key) # Get object with key.
@@ -71,6 +87,38 @@ def delete(hcpm, args):
     else:
         print(f"File: {args.key} does not exist in the HCP")
 
+def search(hcpm, args):
+    if args.query:
+        objects = hcpm.get_objects()
+        found_objs = hcpm.search_objects(args.query)
+        if len(found_objs) > 0:
+            for obj in found_objs:
+                print(obj)
+        else:
+            print(f'No results found for: {args.query}')
+                
+    elif args.qfile:
+        #Read the query file line by line and store in list
+        infile = open(args.qfile, 'r')
+        lines = infile.readlines()
+        #Remove newlines
+        lines = map(lambda s: s.strip(), lines)
+        
+        #Load in all data on HCP
+        objects = hcpm.get_objects()
+        
+        #Search for each item in query file
+        qdict = {}
+        for line in lines:
+            print(f'[-- query: {line} --]')
+            found_objs = hcpm.search_objects(line)
+            if len(found_objs) > 0:
+                for obj in found_objs:
+                    print(obj)
+            else:
+                print('Nothing found')
+    else:
+        print(f'A query or query file needs to be specified if you are using the "search" option')
 
 def arg():
     parser = argparse.ArgumentParser(prog="downloader.py")
@@ -91,7 +139,11 @@ def arg():
     parser.add_argument("--download", action="store_true", help="Use when download")
     parser.add_argument("--delete", action="store_true", help="Use when delete")
     parser.add_argument("--check", action="store_true", help="Only prints results")
-    
+    parser.add_argument("--search", action="store_true", help="Use when searching")
+
+    parser.add_argument("--no-hci", action="store_true", help="Avoid using HCI")
+    parser.add_argument("--qfile", help="Input file of queries")
+
     args = parser.parse_args()
 
     return args
@@ -102,18 +154,23 @@ def main():
     hcpm = HCPManager(args.endpoint, args.aws_access_key_id, args.aws_secret_access_key)
     hcpm.attach_bucket(args.bucket)
 
-    hci.create_template(args)    
-    token = hci.generate_token(args) 
-    pretty = json.loads(hci.query(token, args.index))
+    #Skip the HCI stuff if requested
+    if not args.no_hci:
+        hci.create_template(args)    
+        token = hci.generate_token(args) 
+        pretty = json.loads(hci.query(token, args.index))
 
     if args.download:
-        download(hcpm, args, pretty)
+        download_files(hcpm, args, pretty)
 
     elif args.delete:
         delete(hcpm, args)
     
     elif args.check:
         check(hcpm, args, pretty)
+
+    elif args.search:
+        search(hcpm, args)
 
 if __name__ == "__main__":
     main()
