+++ HCPinterface/cli/legacy_downloader.py
@@ -2,6 +2,7 @@
 
 # Downloads or deletes files from selected bucket on the HCP.
 
+import click
 import glob
 import argparse
 import os
@@ -13,6 +14,12 @@ from hcp.hcp import HCPManager
 from hci import hci
 
 ##############################################
+
+@click.group()
+def legacy_download():
+    """Downloads or deletes files on the old NGS buckets on the HCP"""
+    return
+
 def check(hcpm, args, pretty):
     if args.query:
         try:
@@ -26,9 +33,8 @@ def check(hcpm, args, pretty):
                 lst = string.replace('"','').replace("\\","").replace("[","").replace("]","").replace(";",",").split(",")
             print(f"Metadata file: {meta}")
             for i in lst:
-                if args.query in os.path.basename(i) or args.query in i:
+                if args.query in i:
                     print("check:",i)
-                    name = i.replace(".fastq.gz", ".fasterq").strip() # Replace suffix. 
 
         except:
             print(f"File(s) does not exists: {args.query}")
@@ -44,19 +50,12 @@ def download(hcpm, args, pretty):
         for item in results:
             itm = item["metadata"]
             samples = itm["samples_Fastq_paths"]
-            string = "".join(samples).strip("[]").strip("{]}'")
-            lst = string.replace('"','').replace("\\","").replace("[","").replace("]","").replace(";",",").split(",")
-
-        for i in lst:
-            if args.query in os.path.basename(i) or args.query in i:
-                s = os.path.basename(i)
-                print("downloading:",s.replace(".fastq.gz", ".fasterq").strip())
-                name = s.replace(".fastq.gz", ".fasterq").strip() # Replace suffix. 
-                obj = hcpm.get_object(name) # Get object with json.
-                if obj is not None:
-                    hcpm.download_file(obj, f"{args.output}/"+os.path.basename(name)) # Downloads file.
-                else:
-                    print(f"File: '{name}' does not exist in bucket '{args.bucket}' on the HCP")
+        for i in samples:
+            obj = hcpm.get_object(i) # Get object with json.
+            if obj is not None:
+                hcpm.download_file(obj, f"{args.output}/"+os.path.basename(i)) # Downloads file.
+            else:
+                print(f"File: '{s}' does not exist in bucket '{args.bucket}' on the HCP")
 
     else:
         obj = hcpm.get_object(args.key) # Get object with key.
@@ -79,38 +78,6 @@ def delete(hcpm, args):
     else:
         print(f"File: {args.key} does not exist in the HCP")
 
-def search(hcpm, args):
-    if args.query:
-        objects = hcpm.get_objects()
-        found_objs = hcpm.search_objects(args.query)
-        if len(found_objs) > 0:
-            for obj in found_objs:
-                print(obj)
-        else:
-            print(f'No results found for: {args.query}')
-                
-    elif args.qfile:
-        #Read the query file line by line and store in list
-        infile = open(args.qfile, 'r')
-        lines = infile.readlines()
-        #Remove newlines
-        lines = map(lambda s: s.strip(), lines)
-        
-        #Load in all data on HCP
-        objects = hcpm.get_objects()
-        
-        #Search for each item in query file
-        qdict = {}
-        for line in lines:
-            print(f'[-- query: {line} --]')
-            found_objs = hcpm.search_objects(line)
-            if len(found_objs) > 0:
-                for obj in found_objs:
-                    print(obj)
-            else:
-                print('Nothing found')
-    else:
-        print(f'A query or query file needs to be specified if you are using the "search" option')
 
 def arg():
     parser = argparse.ArgumentParser(prog="downloader.py")
@@ -131,11 +98,7 @@ def arg():
     parser.add_argument("--download", action="store_true", help="Use when download")
     parser.add_argument("--delete", action="store_true", help="Use when delete")
     parser.add_argument("--check", action="store_true", help="Only prints results")
-    parser.add_argument("--search", action="store_true", help="Use when searching")
-
-    parser.add_argument("--no-hci", action="store_true", help="Avoid using HCI")
-    parser.add_argument("--qfile", help="Input file of queries")
-
+    
     args = parser.parse_args()
 
     return args
@@ -146,11 +109,9 @@ def main():
     hcpm = HCPManager(args.endpoint, args.aws_access_key_id, args.aws_secret_access_key)
     hcpm.attach_bucket(args.bucket)
 
-    #Skip the HCI stuff if requested
-    if not args.no_hci:
-        hci.create_template(args)    
-        token = hci.generate_token(args) 
-        pretty = json.loads(hci.query(token, args.index))
+    hci.create_template(args)    
+    token = hci.generate_token(args) 
+    pretty = json.loads(hci.query(token, args.index))
 
     if args.download:
         download(hcpm, args, pretty)
@@ -160,9 +121,6 @@ def main():
     
     elif args.check:
         check(hcpm, args, pretty)
-
-    elif args.search:
-        search(hcpm, args)
 
 if __name__ == "__main__":
     main()
